Переносит функцию отправки данных формы в модуль с формой
